Initial commit of refactored workflow
